+++ src/Command/FilesBackupCommand.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
  */
 namespace FilesBackup\Command;
 
+use Exception;
 use FilesBackup\Command\FilesBackupCommandSubscriber;
 use FilesBackup\FilesBackup;
 use Symfony\Component\Console\Command\Command;
@@ -48,9 +49,9 @@ class FilesBackupCommand extends Command
         $this->setHelp('This command performs a files backup')
             ->addArgument('target', InputArgument::REQUIRED, 'Target zip file you want to create')
             ->addOption('source', 's', InputOption::VALUE_REQUIRED, 'Source directory', defined('APP') ? APP : (defined('ROOT') ? ROOT : getcwd()))
-            ->addOption('exclude', 'e', InputArgument::IS_ARRAY | InputArgument::OPTIONAL, 'Excludes directories from matching. Directories must be relative')
+            ->addOption('exclude', 'e', InputOption::VALUE_IS_ARRAY | InputOption::VALUE_REQUIRED, 'Excludes directories from matching. Directories must be relative')
             ->addOption('no-git-ignore', null, InputOption::VALUE_NONE, 'Does not ignore files and directories specified in the `.gitignore` file')
-            ->addOption('include', 'i', InputArgument::IS_ARRAY | InputArgument::OPTIONAL, 'includes directories excluded by the `.gitignore` file');
+            ->addOption('include', 'i', InputOption::VALUE_IS_ARRAY | InputOption::VALUE_REQUIRED, 'includes directories excluded by the `.gitignore` file');
     }
 
     /**
@@ -95,11 +96,17 @@ class FilesBackupCommand extends Command
 
         try {
             $FilesBackup = new FilesBackup($source, $options ?? []);
-            $FilesBackup->getEventDispatcher()->addSubscriber(new FilesBackupCommandSubscriber($output));
+            $EventDispatcher = $FilesBackup->getEventDispatcher();
+            $EventDispatcher->addSubscriber(new FilesBackupCommandSubscriber($output));
             $FilesBackup->create($target);
 
             $output->writeln('<info>Backup exported successfully to: `' . $target . '`</info>');
-        } catch (\Exception $e) {
+            $fileAddedCount = count($EventDispatcher->getEventList()->extract('FilesBackup.fileAdded'));
+            $output->writeln('<info>File added: ' . $fileAddedCount . '</info>');
+        } catch (Exception $e) {
+            if ($output->isDebug()) {
+                throw $e;
+            }
             $output->writeln('<error>Error: ' . lcfirst($e->getMessage()) . '</error>');
 
             return Command::FAILURE;