+++ tests/TestCase/FilesBackupTest.php
@@ -17,6 +17,7 @@ namespace FilesBackup\Test\TestCase;
 use FilesBackup\FilesBackup;
 use FilesBackup\Test\ZipperReader;
 use FilesBackup\TestSuite\TestCase;
+use Symfony\Component\OptionsResolver\Exception\InvalidOptionsException;
 use Tools\TestSuite\EventAssertTrait;
 use ZipArchive;
 
@@ -36,6 +37,16 @@ class FilesBackupTest extends TestCase
         $file = tempnam(TMP, 'tmp') ?: '';
         $this->expectExceptionMessage('`' . $file . '` is not a directory');
         new FilesBackup($file);
+    }
+
+    /**
+     * Test for `__construct()` method, with a bad option
+     * @test
+     */
+    public function testConstructorBadOption(): void
+    {
+        $this->expectException(InvalidOptionsException::class);
+        new FilesBackup(APP, ['git_ignore' => 'string']);
     }
 
     /**
@@ -107,14 +118,14 @@ class FilesBackupTest extends TestCase
         $this->assertContains(APP . 'logs' . DS . 'error.log', $result);
         $this->assertContains(APP . 'vendor' . DS . 'vendor.php', $result);
 
-        $FilesBackup = new FilesBackup(APP, ['exclude' => 'subDir/subSubDir']);
+        $FilesBackup = new FilesBackup(APP, ['exclude' => 'subDir' . DS . 'subSubDir']);
         $result = $FilesBackup->getAllFiles();
         $this->assertCount($countExpectedFiles - 1, $result);
         $this->assertNotContains(APP . 'logs' . DS . 'error.log', $result);
         $this->assertNotContains(APP . 'vendor' . DS . 'vendor.php', $result);
         $this->assertNotContains(APP . 'subDir' . DS . 'subSubDir' . DS . 'subSubDirFile', $result);
 
-        $FilesBackup = new FilesBackup(APP, ['exclude' => ['subDir/subSubDir', 'subDir/anotherSubDir']]);
+        $FilesBackup = new FilesBackup(APP, ['exclude' => ['subDir' . DS . 'subSubDir', 'subDir' . DS . 'anotherSubDir']]);
         $result = $FilesBackup->getAllFiles();
         $this->assertCount($countExpectedFiles - 2, $result);
         $this->assertNotContains(APP . 'logs' . DS . 'error.log', $result);